Switch to pipeline for file parsing
.pipe has issues with error handling. If there was an error later on, file descriptors would leak and there would eventually be an IO error
